backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
Naked==0.1.32
openai==1.54.3
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.1
status==0.2.5
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
